Added functionality for alternating turns; created person class
I added a variable in "cell" called "token" which checks whether a p1 or p2 (or no) token is present in the cell. Using this and the new player class, I implemented the ability to alternate turns as well as only add tokens when there other tokens below it or or when it is at the bottom.